prepare `calculateBackoff(failedAttempt:)` to be used in `HTTP2StateMaschine` too
